cmd: Remove `mode` command.
